requirements.txt
flask
gtts
google-generativeai
python-dotenv
playsound==1.2.2
pandas